refactor(exceptions): use a custom exception instead of RuntimeExceptions
