+++ src/main/java/com/github/eddranca/datagenerator/visitor/GenerationContext.java
@@ -3,7 +3,7 @@ package com.github.eddranca.datagenerator.visitor;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.eddranca.datagenerator.FilteringBehavior;
-import com.github.eddranca.datagenerator.FilteringException;
+import com.github.eddranca.datagenerator.exception.FilteringException;
 import com.github.eddranca.datagenerator.exception.InvalidReferenceException;
 import com.github.eddranca.datagenerator.generator.FilteringGeneratorAdapter;
 import com.github.eddranca.datagenerator.generator.Generator;
