+++ src/test/java/com/github/eddranca/datagenerator/generator/defaults/ChoiceGeneratorTest.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.github.eddranca.datagenerator.FilteringException;
+import com.github.eddranca.datagenerator.exception.FilteringException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
