+++ src/test/java/com/github/eddranca/datagenerator/visitor/GenerationContextTest.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.github.eddranca.datagenerator.FilteringBehavior;
-import com.github.eddranca.datagenerator.FilteringException;
+import com.github.eddranca.datagenerator.exception.FilteringException;
 import com.github.eddranca.datagenerator.exception.InvalidReferenceException;
 import com.github.eddranca.datagenerator.generator.Generator;
 import com.github.eddranca.datagenerator.generator.GeneratorRegistry;
